add book trip tab completed
