+++ src/components/common/SelectInputField/index.jsx
@@ -1,18 +1,18 @@
-import React from "react";
-import Select from "react-select";
+import React from 'react';
+import Select from 'react-select';
 
 const SelectInputField = props => {
   const {
     options,
     name,
     label,
-    placeholder,
     disabled,
     isSearchable,
     error,
     value,
     onChange,
     wrapperDivClass,
+    iconName,
     ...restProps
   } = props;
 
@@ -20,26 +20,33 @@ const SelectInputField = props => {
     onChange(name, selectedOption);
   };
 
+  const renderIcon = iconName => (
+    <img className="label-icon" src={`./images/icons/${iconName}.svg`} alt="calendar icon" />
+  );
+
   return (
-    <div className={wrapperDivClass}>
-      {label && <label className="vm-form-label">{label}</label>}
-      <div>
-        <Select
-          isClearable={true}
-          openMenuOnClick={true}
-          hideSelectedOptions={true}
-          placeholder={placeholder}
-          className="selectbox-container"
-          classNamePrefix="react-select"
-          name={name}
-          isSearchable={isSearchable || false}
-          options={options}
-          value={value}
-          onChange={handleSelectChange}
-          isDisabled={disabled}
-          {...restProps}
-        />
-      </div>
+    <div className="selectbox-container">
+      {label && (
+        <label className="vm-form-label">
+          {iconName && renderIcon(iconName)}
+          {label}
+        </label>
+      )}
+      <Select
+        isClearable={true}
+        openMenuOnClick={true}
+        hideSelectedOptions={true}
+        placeholder="Choose..."
+        className="selectbox"
+        classNamePrefix="react-select"
+        name={name}
+        isSearchable={isSearchable || false}
+        options={options}
+        value={value}
+        onChange={handleSelectChange}
+        isDisabled={disabled}
+        {...restProps}
+      />
       {error && <div className="invalid-feedback d-block">{error}</div>}
     </div>
   );
